Improve rendering performance
Previously there were a ton of draw calls used since every image and
every text was rendered to its own texture. However since we always
display the same icons, we only need to render everything once per
resize.

To improve performance when the application is already open, everything
is now rendered to a single texture and rendered from there over and
over again.

Since there's no need to recognize installation while the application is
open, the icons themselves are stored permanently, which should help in
case the application is resized.