Add support for multiple textures
Rendering everything in a single texture is nice when you only have
about a screen full of data to display, but with bigger texture sizes
devices will stop working properly.

To make sure Tzompantli works regardless of how bloated the system is,
there is now a mechanism that splits all the applications into multiple
textures and then stiches them back together with multiple draw calls
during rendering.

The value of `MAX_TEXTURE_ROWS` specifies how many rows are stored in
each texture. A value that is too big will cause nothing to be rendered,
however during testing with a value of `10` it was also noticeable that
the quality of the texture reduced the further away from the origin you
scroll.

The value `5` was picked because that didn't show any quality loss over
the entire duration of the texture with my PinePhone.