Improve frame scheduling
Instead of requesting a frame whenever rendering is done, a new frame is
now only requested when the user scrolls through the list of
applications and there is not a frame request pending already.

This should basically reduce the number of redraws to 0 while the
application is not interacted with.

Fixes #4.