Fix rendering without any icons
Due to a clippy suggestion, a `.min.max` was incorrectly changed to a
`.clamp` in b3ea3e8, this caused a crash with empty icons since the
lower bound must be `<=` than the upper bound for `clamp`.

The `icon_count` has also been clamped with `.max(2)`, since there will
always be at least 2 icons for the power symbols.

Closes #17.