Sort desktop entries
The initial desktop entry storage was changed to a `HashMap` when the
desktop entry fallback was implemented. As a result, this caused the
order of desktop entries to be completely randomized.

To make the order of elements obvious to the user and consistent, all
applications are now listed in alphabetical order.