Update the SVG crates
The current version was from February 2022, many fixes and improvements
happened since then.

This fixes the rendering of the icons of adwaita-1-demo, audacity and
winetricks on my system, the latter two using blur.  This also makes
gaussian blur take the most CPU time out of the rendering, but with
other optimisations this is actually a net win.

On my PinePhone, the total time it takes to display the first frame goes
from approximately 16s down to ~1.8s, from a very unscientific use of
`time` averaged multiple times.  On a flamegraph, we can see the vast
majority of that time was spent loading fonts before (79.5%), while this
is not done at all now.  According to resvg’s ChangeLog, since 0.28 usvg
doesn’t convert text into paths any longer by default.