Revive integer scale and make wp_fractional_scale optional
Weston doesn’t implement it yet, and this makes it easier to test it on
my laptop.  This protocol is inherently racy, as we don’t know on which
output we are before we send the first buffer, but oh well.  Let’s just
wait until all compositors have implemented wp_fractional_scale to
remove it again this time.

This made me aware that we pick the default size of 64×64 statically,
while scaling changes will make this decision suboptimal, we might want
to do something about that issue someday.