Delay EGL surface init
When the compositor has an opinion on the size we should have, it is
better to adopt it immediately, as otherwise we will waste the initial
draw and resize and draw again. And besides being a waste of resources,
it is also visually unpleasant to see a short frame at the wrong
resolution.