Optimize premultiply algorithm
Switch the premultiply algorithm from f32 to u16, using SIMD on AArch64.

This makes IconLoader::load() go down from 3.36% to 0.25% of the total profile
on my Cortex-A53. For reference the non-SIMD but u16-based function takes
1.46%. On my laptop it goes from 2.62% down to 0.89% with the generic path.

The AArch64 code makes use of the LD4/ST4 instructions, which allow us
to work on 8 pixels at once instead of 1.