Split SVG rendering in two steps
resvg has the interesting property that it does half of the work during
Tree building, to then make rendering as fast as possible.

For now we only use this to avoid re-parsing the poweroff and reboot
icons, but in the future we can use this also on other icons.

In the future we will also do the parsing early in a thread pool, and
only do the rendering once we know the size at which we will get
displayed.