Move config to shared module
This moves all configuration values which might be interesting for users
to a shared `config` module, making it easier to share these constants
between modules and change them in a single place.

This also changes the colors around a bit, since the default background
color was previously `0.1` which doesn't map well to a u8 (`25.5`).
Instead it uses `24` (`#181818`) as u8 color value, to ensure there's
never any rounding issues.