Add icon theme support
This fixes the icon load order to parse the `index.theme` from the
`default` theme by default, then following along the `Inherits`
attributes to construct a full theme fallback list.

Icons are always only loaded from one theme, but theme fallback is used
in case an icon is not present in a higher-level icon theme.

Additionally a bug with symbolic themes has been fixed, which
incorrectly stripped `-symbolic` as a prefix, when it should have been
stripped as a suffix.

Closes #9.