Add config icon for hiding desktop entries
This adds a cog between poweroff and reboot which enters a separate
mode in which individual applications can be hidden or shown by tapping
on their icon.

Icons are hidden by adding a desktop file with the content
`NoDisplay=true` to the user's data directory, or editing the existing
desktop entry by appending this option or replacing an existing
`NoDisplay=false`.

To show icons, all desktop files with `NoDisplay=true` are edited to use
`NoDisplay=false` instead, removing the file if that was the only
content. If writing to one of these files fails, the desktop entry is
removed completely from Tzompantli, since the user does not have the
permissions to make it visible without creating a full desktop file.

Closes #18.