Replace renderer with Skia
This is a full rewrite of the renderer to replace it with a Skia
renderer that has better text support and renders the individual icons
on-demand rather than rasterizing all of them ahead of time.

While this should perform better than the previous renderer, especially
with a large number of icons, it does still miss frames when loading new
icons on slow hardware. Part of this seems to be due to performance
issues with Skia.

One notable Skia performance issue is the SVG rasterization speed. This
patch uses resvg to draw all the SVGs into CPU buffers in parallel,
rather than using Skia's SVG renderer, since the Skia renderer is slower
than resvg even if the latter is not running parallel. The parallel SVG
rasterization benefits are marginal on the PinePhone, though could bring
larger benefits on stronger hardware.

Closes #41.